+++ rakaia/components/layout.py
@@ -1593,10 +1593,30 @@ def register_app_layout(config: dict, cache_dest: Union[str, Path]):
                                 columns=[], data=None, editable=False, filter_action='native',
                                 style_table={"max-width": "inherit", "overflowX": "auto"})]),
                                 id="show-clust-dist-table", size='l'),
+                                html.Br(),
+                                html.Br(),
+                                dbc.Modal(children=dbc.ModalBody([
+                                html.Div([
+                                html.B("Differential gene expression for overlay"),
+                                dbc.Button("Add selected marker to canvas", id="dge-to-canvas",
+                                style={"background-color": DEFAULT_WIDGET_COLOUR, "margin-right": '7.5px',
+                                       'float': "right", 'margin-left': '10px'}),
+                                dbc.Tooltip(TOOLTIPS['dge-show'], target="dge-to-canvas",
+                                style={"display": "flex", "justifyContent": "center"}, placement='right'),
+                                ], style={"display": "flex", "justify-content": "space-evenly", "width": "75%"}),
+                                html.Br(),
+                                dash_table.DataTable(id='dge-table', sort_action="none", filter_action="none",
+                                columns=[], data=None, editable=False, style_table={
+                                "max-width": "inherit", "overflowX": "auto"}),
+                                ]), id="show-dge-table", size='xl'),
                                 dbc.Collapse(html.Div([
                                 html.Br(),
                                 html.H6(children=[], id="cluster-assignments")]),
                                              is_open=False, id='cluster-label-collapse'),
+                                dbc.Button("Differential gene expression for overlay", id="dge-overlay-show",
+                                    style={"background-color": DEFAULT_WIDGET_COLOUR, "margin-right": '7.5px'}),
+                                dbc.Tooltip(TOOLTIPS['dge'], target="dge-overlay-show",
+                                    style={"display": "flex", "justifyContent": "center"}, placement='left'),
                                 dbc.Modal(id="quantification-roi-modal", children=dbc.ModalBody([
                                 html.Div([dbc.Button("Quantify current ROI", id="quantify-cur-roi-execute",
                                            style={"background-color": DEFAULT_WIDGET_COLOUR, "margin-right": '7.5px'}),