+++ rakaia/utils/alert.py
@@ -172,7 +172,10 @@ class ToolTips(BaseModel):
                                               " to set the position for adding the canvas blend to a stitched image.",
                       "stitch-coord-min": "This is the top-left coordinate in the stitch to anchor the placement of a new blended image.",
                       "wsi-zoom-level": "0 is fully zoomed out, while 1 is the most zoomed in. Use log scale for "
-                                        "zooming proportional to magnification, or disable for equal zoom increments."}
+                                        "zooming proportional to magnification, or disable for equal zoom increments.",
+                      "dge": "Show differential expression analysis on the current ROI with the top 25 "
+                             "markers per category in the current overlay. Works only on ROIs from an Anndata or zarr store.",
+                      "dge-show": "Note: only a single marker is added at a time (the cell with a red outline)."}
 
 
 class PanelMismatchError(Exception):