+++ CHANGELOG.md
@@ -7,17 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.30.0] - 2026-03-30
+## [0.30.0] - 2026-04-09
 
 ## Added
 
 - Transfer stitched image to WSI view (low-res preview remains)
 - Updated Visium HD export with masks in `examples`
+- Modal for differential gene expression for Anndata ROIs with categorical overlay
 
 ## Changed
 
 - Add better permission on one-directory standalone builds
-- Update `dependabot` vulnerabilities: `requests`, `crytography`
+- Update `dependabot` vulnerabilities: `requests`, `crytography`, `Flask-HTTPAuth`
 - Cache mask boundaries when first created for ROIs that are sufficiently large
 - Set dataset gallery down-sampling default to `True`